Start moving Magnet type and utility functions into one file
I needed to refer to the type on both client and server, so having it defined inside
page.client.ts was getting problematic.

I've also refactored the <PrintAllGrids> thing into a `setupMagnetGroups()` function that just
works on the data, not on React components
