+++ app/components/MagnetGrid.tsx
@@ -1,10 +1,6 @@
-import { Magnet } from "~/store/page.client";
+import { SignedMagnet } from "~/store/Magnet";
 
-export type MagnetImage = Magnet & {
-  presignedUrl: string;
-};
-
-export function MagnetGrid({ images }: { images: Array<MagnetImage> }) {
+export function MagnetGrid({ images }: { images: Array<SignedMagnet> }) {
   const noMoreThan9Images = images.slice(0, 9);
   return (
     <ul className="grid grid-cols-3 grid-rows-3 place-content-around aspect-square w-[125mm] gap-[6mm] p-[3mm] m-3 bg-white border-2">
